Move `checkInterfaceAddress()` into `InterfaceAddress`.
